+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Field `collector_url` added to the `jaeger` tracer.
 - The bloblang method `strip_html` now allows you to specify a list of allowed elements.
 - New bloblang method `parse_xml`.
+- New bloblang method `replace_many`.
 
 ## 3.37.0 - 2021-01-06